always trigger manual text events
This code was originally based on fixing the first character input beep, but since there appears to be no side effect of manually creating all input events this would fix #97. This other solution would be to extend the `window.getSelection()` check to be something like `(window.getSelection().baseOffset === 0 || window.getSelection().focusOffset === 0)` but it’s starting to look fragile.