Add Paste Image Support
Since WebView is weird and doesn't support pasting image data directly (the javascript file dataTransfer event is blank), and you can't really fake a dataTransfer event, one would try pasting an image but it wouldn't do anything.

However, since the fact that drag + drop does work (yay!), we can synthesize a fake drag + drop session anytime someone attempts to paste an image. If the image is on the local filesystem it'll just upload whatever it's given, but if it's clipboard data, it'll convert it to PNG and then upload that.

If the pasteboard just contains text (or anything other than an image), original pasting functionality is performed.

Yay for fake drag + drop!