Adds config for hiding traffic light and controlling autoreload
Defaults key "autoreload" can be set to a positive fractional number-of-seconds to enable automatic reloading of messenger.com while the app is not active (screensaver running, screen locked, app hidden, app in background, app minimized, etc). A zero or negative value, or absence of a value causes autoreload to be disabled.

Defaults key "traffic-lights/disable" can be set to boolean YES to hide the window controls (close, minimize, zoom).